op2:  - adding nx 2020.2  - adding RealCompositePlateStrengthRatioArray with write_f06
