+++ pyNastran/op2/op2_interface/nx_tables.py
@@ -9,7 +9,7 @@ NX_VERSIONS = [
     '11.0', '11.0.1', '11.0.2',
     '12.0', '12.0.2',
     '2019.0', '2019.1', '2019.2',
-    '2020.0', '2020.1',
+    '2020.0', '2020.1', '2020.2',
 ]
 
 NX_ELEMENTS = {
