+++ pyNastran/op2/tables/oug/oug_velocities.py
@@ -13,7 +13,7 @@ class RealVelocityArray(RealTableArray):
         words = ['                                                   V E L O C I T Y   V E C T O R\n', ]
         #' \n',
         #'      POINT ID.   TYPE          T1             T2             T3             R1             R2             R3\n']
-        if self.table_name in ['OUGV1', 'OUGV2', 'OUPV1', 'BOUGV1']:
+        if self.table_name in ['OUGV1', 'OUGV2', 'OUPV1', 'BOUGV1', 'OVG1']:
             pass
         elif self.table_name in ['OUXY1', 'OUXY2']:
             words = ['                                           V E L O C I T Y   V E C T O R   (SOLUTION SET)']
